requirements.txt
gym
tensorflow==1.*
baselines
dataclasses
scikit-image
PyGeometry
pyyaml
pyglet
stable-baselines3[extras]